backend/requirements.txt
Flask
google-generativeai
python-dotenv
pandas
scikit-learn
supabase